refactor(logs): move sanitize
